move manifest artifacts to dbt/artifacts
