+++ core/dbt/artifacts/manifest.py
@@ -1,0 +1,175 @@
+from dataclasses import dataclass, field
+from typing import Mapping, Iterable, Tuple, Optional, Dict, List, Any
+from uuid import UUID
+
+from dbt.artifacts.base import (
+    BaseArtifactMetadata,
+    ArtifactMixin,
+    schema_version,
+    get_artifact_schema_version,
+)
+
+
+# TODO: remove usage of dbt modules other than dbt.artifacts
+from dbt.contracts.graph.manifest_upgrade import upgrade_manifest_json
+from dbt import tracking
+from dbt.flags import get_flags
+from dbt.contracts.graph.nodes import (
+    Documentation,
+    Exposure,
+    GraphMemberNode,
+    Group,
+    Macro,
+    ManifestNode,
+    Metric,
+    SavedQuery,
+    SemanticModel,
+    SourceDefinition,
+)
+
+
+NodeEdgeMap = Dict[str, List[str]]
+UniqueID = str
+
+
+@dataclass
+class ManifestMetadata(BaseArtifactMetadata):
+    """Metadata for the manifest."""
+
+    dbt_schema_version: str = field(
+        default_factory=lambda: str(WritableManifest.dbt_schema_version)
+    )
+    project_name: Optional[str] = field(
+        default=None,
+        metadata={
+            "description": "Name of the root project",
+        },
+    )
+    project_id: Optional[str] = field(
+        default=None,
+        metadata={
+            "description": "A unique identifier for the project, hashed from the project name",
+        },
+    )
+    user_id: Optional[UUID] = field(
+        default=None,
+        metadata={
+            "description": "A unique identifier for the user",
+        },
+    )
+    send_anonymous_usage_stats: Optional[bool] = field(
+        default=None,
+        metadata=dict(
+            description=("Whether dbt is configured to send anonymous usage statistics")
+        ),
+    )
+    adapter_type: Optional[str] = field(
+        default=None,
+        metadata=dict(description="The type name of the adapter"),
+    )
+
+    def __post_init__(self):
+        if tracking.active_user is None:
+            return
+
+        if self.user_id is None:
+            self.user_id = tracking.active_user.id
+
+        if self.send_anonymous_usage_stats is None:
+            self.send_anonymous_usage_stats = get_flags().SEND_ANONYMOUS_USAGE_STATS
+
+    @classmethod
+    def default(cls):
+        return cls(
+            dbt_schema_version=str(WritableManifest.dbt_schema_version),
+        )
+
+
+@dataclass
+@schema_version("manifest", 12)
+class WritableManifest(ArtifactMixin):
+    nodes: Mapping[UniqueID, ManifestNode] = field(
+        metadata=dict(description=("The nodes defined in the dbt project and its dependencies"))
+    )
+    sources: Mapping[UniqueID, SourceDefinition] = field(
+        metadata=dict(description=("The sources defined in the dbt project and its dependencies"))
+    )
+    macros: Mapping[UniqueID, Macro] = field(
+        metadata=dict(description=("The macros defined in the dbt project and its dependencies"))
+    )
+    docs: Mapping[UniqueID, Documentation] = field(
+        metadata=dict(description=("The docs defined in the dbt project and its dependencies"))
+    )
+    exposures: Mapping[UniqueID, Exposure] = field(
+        metadata=dict(
+            description=("The exposures defined in the dbt project and its dependencies")
+        )
+    )
+    metrics: Mapping[UniqueID, Metric] = field(
+        metadata=dict(description=("The metrics defined in the dbt project and its dependencies"))
+    )
+    groups: Mapping[UniqueID, Group] = field(
+        metadata=dict(description=("The groups defined in the dbt project"))
+    )
+    selectors: Mapping[UniqueID, Any] = field(
+        metadata=dict(description=("The selectors defined in selectors.yml"))
+    )
+    disabled: Optional[Mapping[UniqueID, List[GraphMemberNode]]] = field(
+        metadata=dict(description="A mapping of the disabled nodes in the target")
+    )
+    parent_map: Optional[NodeEdgeMap] = field(
+        metadata=dict(
+            description="A mapping from child nodes to their dependencies",
+        )
+    )
+    child_map: Optional[NodeEdgeMap] = field(
+        metadata=dict(
+            description="A mapping from parent nodes to their dependents",
+        )
+    )
+    group_map: Optional[NodeEdgeMap] = field(
+        metadata=dict(
+            description="A mapping from group names to their nodes",
+        )
+    )
+    saved_queries: Mapping[UniqueID, SavedQuery] = field(
+        metadata=dict(description=("The saved queries defined in the dbt project"))
+    )
+    semantic_models: Mapping[UniqueID, SemanticModel] = field(
+        metadata=dict(description=("The semantic models defined in the dbt project"))
+    )
+    metadata: ManifestMetadata = field(
+        metadata=dict(
+            description="Metadata about the manifest",
+        )
+    )
+
+    @classmethod
+    def compatible_previous_versions(cls) -> Iterable[Tuple[str, int]]:
+        return [
+            ("manifest", 4),
+            ("manifest", 5),
+            ("manifest", 6),
+            ("manifest", 7),
+            ("manifest", 8),
+            ("manifest", 9),
+            ("manifest", 10),
+            ("manifest", 11),
+        ]
+
+    @classmethod
+    def upgrade_schema_version(cls, data):
+        """This overrides the "upgrade_schema_version" call in VersionedSchema (via
+        ArtifactMixin) to modify the dictionary passed in from earlier versions of the manifest."""
+        manifest_schema_version = get_artifact_schema_version(data)
+        if manifest_schema_version <= 10:
+            data = upgrade_manifest_json(data, manifest_schema_version)
+        return cls.from_dict(data)
+
+    def __post_serialize__(self, dct):
+        for unique_id, node in dct["nodes"].items():
+            if "config_call_dict" in node:
+                del node["config_call_dict"]
+            if "defer_relation" in node:
+                del node["defer_relation"]
+        return dct
